Bring console helpers back into metta/setup
