+++ metta/setup/components/observatory_fe.py
@@ -1,6 +1,6 @@
-from metta.common.util.console_messages import info, success, warning
 from metta.setup.components.base import SetupModule
 from metta.setup.registry import register_module
+from metta.setup.utils import info, success, warning
 
 
 @register_module
