+++ metta/setup/components/observatory_key.py
@@ -1,9 +1,9 @@
 import subprocess
 
-from metta.common.util.console_messages import info, success, warning
 from metta.common.util.stats_client_cfg import get_machine_token
 from metta.setup.components.base import SetupModule
 from metta.setup.registry import register_module
+from metta.setup.utils import info, success, warning
 
 
 @register_module
