+++ metta/setup/components/skypilot.py
@@ -1,8 +1,8 @@
 import subprocess
 
-from metta.common.util.console_messages import info, success
 from metta.setup.components.base import SetupModule
 from metta.setup.registry import register_module
+from metta.setup.utils import info, success
 
 
 @register_module
