+++ metta/setup/components/tailscale.py
@@ -2,10 +2,10 @@ import json
 import platform
 import subprocess
 
-from metta.common.util.console_messages import info, success, warning
 from metta.setup.components.base import SetupModule
 from metta.setup.config import UserType
 from metta.setup.registry import register_module
+from metta.setup.utils import info, success, warning
 
 
 @register_module
