+++ metta/setup/components/wandb.py
@@ -1,9 +1,9 @@
 import os
 import subprocess
 
-from metta.common.util.console_messages import info, success, warning
 from metta.setup.components.base import SetupModule
 from metta.setup.registry import register_module
+from metta.setup.utils import info, success, warning
 
 
 @register_module
